rate movie done

+++ 211283E_Advanced/app/src/main/java/com/example/a211283e_advanced/MovieModel.kt
@@ -12,7 +12,7 @@ data class MovieModel (
     var isViolence:Boolean,
     var isLanguageUsed:Boolean,
     // Allow null
-    var rating:Int?,
+    var rating:Float?,
     var review:String?
     ){
     companion object{
